requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
PyPDF2==3.0.1
langchain==0.0.350
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1 